requirements.txt
flask==3.0.3
Pillow==10.4.0
rembg==2.0.61
gunicorn==23.0.0